pyproject.toml
[tool.poetry]
name = "ai-news-agent"
version = "0.1.0"
description = ""
authors = ["tmelwig <tmelwig@gmail.com>"]
packages = [{include = "ai_news_agent", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
feedparser = ">=6.0.12,<7.0.0"
python-dotenv = ">=1.2.2,<2.0.0"
newspaper3k = "^0.2.8"

[tool.poetry.scripts]
run-agent = "ai_news_agent.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
